CarCondition/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
Pillow
numpy
torch
torchvision
python-dotenv